+++ booksnake/__init__.py
@@ -18,17 +18,11 @@ from booksnake.printing import *
 from booksnake.searchers import *
 from booksnake.sending import *
 
-__version__ = "0.2.1"
-print("October 24, 2017")
 
-# Uniformize 'input()'
-try:
-    input = raw_input
-except NameError:
-    pass
+__version__ = "0.2.2"
 
-settings = {}
-cleanups = []
+SETTINGS = {}
+CLEANUPS = []
 HANDLED_FILETYPES = ['mobi', 'html', 'azw']
 
 
@@ -45,15 +39,15 @@ def read_settings():
         None
 
     """
-    global settings
+    global SETTINGS
     try:
         # Open the settings file
         with open(os.path.expanduser('~/.booksnakerc'), 'r') as setfh:
-            settings = json.load(setfh)
+            SETTINGS = json.load(setfh)
     except Exception as exc:
         # Indicate that a reconcilable failure occurred:
         pretty_print([YELLOW], "No ~/.booksnakerc file found, creating...")
-        settings = {}
+        SETTINGS = {}
         save_settings()
 
 
@@ -71,7 +65,7 @@ def save_settings():
     with open(os.path.expanduser('~/.booksnakerc'), 'w+') as setfh:
         # Write the settings to disk:
         json.dump(
-            settings, setfh, sort_keys=True, indent=4, ensure_ascii=False
+            SETTINGS, setfh, sort_keys=True, indent=4, ensure_ascii=False
         )
 
 
@@ -92,11 +86,11 @@ def convert_file(filename):
     os.system('kindlegen "{}"'.format(filename))
 
     # Add the current filename to the list of things to delete:
-    cleanups.append(filename)
+    CLEANUPS.append(filename)
 
     # "Guess" the name that the converted file will have:
     filename = ".".join(filename.split('.')[:-1]) + ".mobi"
-    cleanups.append(filename)
+    CLEANUPS.append(filename)
     return filename
 
 
@@ -269,7 +263,6 @@ def process_query(query, modes=[], chooser=cli_chooser):
 
     options = []
     for s in modes:
-        options += s.get_options(query)
         try:
             # This fails in cases where the website is down or the HTML has
             # changed dramatically:
@@ -282,7 +275,6 @@ def process_query(query, modes=[], chooser=cli_chooser):
     selection = chooser(options)
     choice = options[int(selection) - 1]
     return _attempt_url(
-        # choice[3].replace('download', 'get'), choice[2], choice[1]
         choice.url, choice.fmt, choice.title
     )
 
@@ -293,7 +285,7 @@ def delete_files():
 
     If a deletion fails, does not notify the user.
     """
-    for fname in cleanups:
+    for fname in CLEANUPS:
         try:
             os.remove(fname)
         except OSError as exp:
@@ -307,7 +299,7 @@ def main():
     Args:
         None
     """
-    global cleanups
+    global CLEANUPS
     read_settings()
     parser = argparse.ArgumentParser(
         description='Search and send books to Kindle.'
@@ -390,7 +382,7 @@ def main():
         searchers = []
         if sum([
             args.use_gutenberg,
-            settings.get('searchers.gutenberg', False)
+            SETTINGS.get('searchers.gutenberg', False)
         ]):
             searchers.append(GutenbergSearcher())
         searchers.append(LibreLibSearcher())
@@ -403,8 +395,8 @@ def main():
     # `filename` holds the name of the file to send.
 
     if args.send_file:
-        cleanups += send_file(filename, args.from_email, args.to_email,
-                              settings=settings)
+        CLEANUPS += send_file(filename, args.from_email, args.to_email,
+                              settings=SETTINGS)
 
     if not args.keep_file:
         delete_files()